move util content to first.c
